+++ functions/package.json
@@ -14,10 +14,13 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
     "firebase-admin": "^9.4.1",
     "firebase-functions": "^3.11.0"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.8",
     "typescript": "4.1.2"
   }
 }
