+++ functions/src/index.ts
@@ -3,8 +3,8 @@ import { initializeApp } from 'firebase-admin';
 initializeApp();
 
 export { addTeamMember } from './addTeamMember';
+export { api } from './api';
 export { createTeam } from './createTeam';
-export { current } from './getCurrentPairs';
 export { removeTeamMember } from './removeTeamMember';
 export { saveHistory } from './saveHistory';
 export { updateTeamSettings } from './updateTeamSettings';
